requirements.txt
django
djangorestframework
psycopg2
pandas
jinja2
